LayoutTests/http/tests/security/location-cross-origin.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<iframe src="http://localhost:8000/security/resources/iframe-with-element.html"></iframe>
<script>
description("Test security checking for access to Location.");
jsTestIsAsync = true;

onload = function() {
    shouldBeUndefined("frames[0].location.protocol");
    shouldBeUndefined("frames[0].location.host");
    shouldBeUndefined("frames[0].location.hostname");
    shouldBeUndefined("frames[0].location.port");
    shouldBeUndefined("frames[0].location.pathname");
    shouldBeUndefined("frames[0].location.search");
    shouldBeUndefined("frames[0].location.hash");
    shouldBeUndefined("frames[0].location.origin");
    shouldBeUndefined("frames[0].location.ancestorOrigins");
    shouldThrow("frames[0].location.toString()");
    shouldThrow("frames[0].location.reload()");
    shouldThrow("frames[0].location.assign('about:blank')");
    // The specification seems to allow access to href but Firefox does not.
    shouldBeUndefined("frames[0].location.href");

    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'protocol').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'host').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'hostname').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'port').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'pathname').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'search').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'hash').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'origin').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'ancestorOrigins').get.call(frames[0].location)");
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'toString').value.call(frames[0].location)");
    // The specification seems to allow access to href but Firefox does not.
    shouldBeUndefined("Object.getOwnPropertyDescriptor(window.location, 'href').get.call(frames[0].location)");

    finishJSTest();
};
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
